Edit profile page completed
