+++ kegeland-frontend/src/pages/profile/EditProfilePage.tsx
@@ -4,109 +4,84 @@ import Card from '../../components/Card';
 import withLayout from '../../hoc/withLayout';
 import withSpinner from '../../hoc/withSpinner';
 import useWheelchairPatient from '../../hooks/useWheelchairPatient';
-import useAuth from '../../hooks/useAuth';
 import { useSelector } from 'react-redux';
 import { RootState } from '../../state/store';
 import { Input } from '@chakra-ui/react';
 import { useEffect, useState } from 'react';
+import { UpdateWheelchairPatientData } from '../../state/ducks/wheelchairPatients/wheelchairPatients.interface';
 
-type Gender = 'M' | 'F' | 'O';
-
-type FormData = {
-  firstName: string;
-  lastName: string;
-  gender: Gender;
-  birthofdate: Date;
-  height: number;
-  weight: number;
-};
 
 const EditProfilePage = () => {
 
     const { authUser } = useSelector((state: RootState) => state.auth);
 
-    const { wheelchairPatient, loading } = useWheelchairPatient(authUser?.id);
 
-    const { userDetails } = useAuth();
+    const { wheelchairPatient, loading, updatePatientData } = useWheelchairPatient(authUser?.id);
 
     const navigate = useNavigate();
 
-    const update = (data: FormData) => {
-      // TODO: Implement update logic
-      // Send data to API
-      // navigate to '/myprofile';
-      console.log(data);
-    };
-
     useEffect(() => {
       if (wheelchairPatient) {
-        const { gender, currentPhysicalState } = wheelchairPatient;
-          setFormData({
-              ...formData,
-              gender: gender || 'M',
-              birthofdate: new Date(),
-              height: currentPhysicalState.height || 0,
-              weight: currentPhysicalState.weight || 0,
-          });
+        setUpdatedData({
+          pid: wheelchairPatient.id,
+          currentPhysicalState: {
+            ...wheelchairPatient.currentPhysicalState, // Preserve other values
+            height: wheelchairPatient.currentPhysicalState.height || 0,
+            weight: wheelchairPatient.currentPhysicalState.weight || 0,
+          },
+        });
       }
     }, [wheelchairPatient]);
-
-
-    useEffect(() => {
-      if (userDetails) {
-        const { firstName, lastName } = userDetails.name;
-        setFormData({
-          ...formData,
-          firstName: firstName || '',
-          lastName: lastName || '',
-        });
-      }
-    }, [userDetails]);
-    
   
-    const [formData, setFormData] = useState<FormData>({
-      firstName: '',
-      lastName: '',
-      gender: 'M' as Gender,
-      birthofdate: new Date(),
-      height: 0,
-      weight: 0,
+    const [updatedData, setUpdatedData] = useState<UpdateWheelchairPatientData>({
+      pid: wheelchairPatient?.id ? wheelchairPatient.id : '',
+      currentPhysicalState: {
+        height: 0,
+        weight: 0,
+        maxHeartRate: 0,
+        averageHeartRate: 0,
+        maxWheelchairSpeed: 0,
+        averageWheelchairSpeed: 0,
+      },
     });
+
+    const update = () => {
+      updatePatientData(updatedData);
+      navigate('/myprofile');
+    };
+
   
-    const handleChange = (e: React.ChangeEvent<HTMLInputElement | HTMLSelectElement>) => {
-      if (e.target.name === "birthofdate") {
-        const newDate = new Date(e.target.value);
-        setFormData({
-          ...formData,
-          birthofdate: newDate
-        });
-      } else {
-        setFormData({
-          ...formData,
-          [e.target.name]: e.target.value,
-        });
-      }
+    const handleHeightChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+      const height = parseFloat(e.target.value);
+      setUpdatedData((prevData) => ({
+        ...prevData,
+        currentPhysicalState: {
+          ...prevData.currentPhysicalState,
+          height,
+        },
+      }));
     };
 
-    // Denne må oppdateres med alle feltene som skal sjekkes (Date og Select)
+    const handleWeightChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+      const weight = parseFloat(e.target.value);
+      setUpdatedData((prevData) => ({
+        ...prevData,
+        currentPhysicalState: {
+          ...prevData.currentPhysicalState,
+          weight,
+        },
+      }));
+    };  
+    
+
+    
     const allFilled = (): boolean => {
-      return Object.values(formData).every(field => {
-        if (typeof field === 'string') {
-          return field.trim() !== ''; // Check if string fields are not empty
-        }
-        if (typeof field === 'number') {
-          return field !== 0; // Assume 0 is not a valid value. Adjust if not the case.
-        }
-        return true; // Assumes other fields (like Date) are always valid
-      });
+      return (
+        updatedData.currentPhysicalState.height > 0 &&
+        updatedData.currentPhysicalState.weight > 0
+      );
     };
-
-    const formatDate = (date: Date) => {
-      const yyyy = date.getFullYear().toString();
-      const mm = (date.getMonth() + 1).toString().padStart(2, '0');
-      const dd = date.getDate().toString().padStart(2, '0');
-      return `${yyyy}-${mm}-${dd}`;
-    };
+    
     
   
     return (
@@ -126,41 +101,17 @@ const EditProfilePage = () => {
           <div style={{marginLeft: '20%', marginRight: '20%'}}>
             <Stack spacing={10} direction="row" mt={6}>
               <FormControl>
-                <FormLabel>First Name</FormLabel>
-                <Input name="firstName" value={formData.firstName} onChange={handleChange} />
-              </FormControl>
-              <FormControl>
-                <FormLabel>Last Name</FormLabel>
-                <Input name="lastName" value={formData.lastName} onChange={handleChange} />
-              </FormControl>
-            </Stack>
-            <Stack spacing={10} direction="row" mt={6}>
-              <FormControl id="gender">
-                <FormLabel>Gender</FormLabel>
-                <Select name="gender" value={formData.gender} onChange={handleChange}>
-                  <option value="M">Male</option>
-                  <option value="F">Female</option>
-                  <option value="O">Other</option>
-                </Select>
-              </FormControl>
-              <FormControl>
-                <FormLabel>Birth of Date</FormLabel>
-                <Input name="birthofdate" type="date" value={formatDate(formData.birthofdate)} onChange={handleChange} />
-              </FormControl>
-            </Stack>
-            <Stack spacing={10} direction="row" mt={6}>
-              <FormControl>
                 <FormLabel>Height</FormLabel>
-                <Input name="height" type='number' value={formData.height} onChange={handleChange} />
+                <Input name="height" type='number' value={updatedData.currentPhysicalState.height} onChange={handleHeightChange} />
               </FormControl>
               <FormControl>
                 <FormLabel>Weight</FormLabel>
-                <Input name="weight" value={formData.weight} onChange={handleChange} />
+                <Input name="weight" type='number' value={updatedData.currentPhysicalState.weight} onChange={handleWeightChange} />
               </FormControl>
             </Stack>
             <Stack direction="row" spacing={10} mt={10} justifyContent="center">
               <Button colorScheme="gray" onClick={() => navigate('/myprofile')}>Cancel</Button>
-              <Button colorScheme="blue" disabled={!allFilled()} onClick={() => update(formData)}>Submit</Button>
+              <Button colorScheme="blue" disabled={!allFilled()} onClick={update}>Submit</Button>
             </Stack>
           </div>
         </Card>
